requirements.txt
botbuilder-integration-aiohttp>=4.14.0
openai
openai[datalib]
azure-core
azure-search-documents==11.4.0b8